requirements.txt
# python version 3.7
# pip install -r requirements.txt

django
djangorestframework
djangorestframework-simplejwt
django-environ
python-dotenv
pillow
django-cors-headers
gunicorn